Use the --copy-files option with babel
In case you are using file loaders or css/sass/scss/less/etc loaders with your webpack and you want to use this component in a project that also has those loaders, you have to copy those assets to the `dist` directory as well, so the webpack on the parent project can find them and compile the bundle. If you don't have any of those using the option will not hurt.